Cast MaxIter solver option params to unsigned int

+++ test/integration/test_hform_rod.cpp
@@ -85,7 +85,7 @@ protected:
 
     hephaestus::InputParameters solver_options;
     solver_options.SetParam("Tolerance", float(1.0e-16));
-    solver_options.SetParam("MaxIter", 1000);
+    solver_options.SetParam("MaxIter", (unsigned int)1000);
     solver_options.SetParam("PrintLevel", 0);
 
     hephaestus::Variables variables;
